requirements.txt
Django==1.8
Pygments==2.0.2
argparse==1.2.1
blessings==1.6
bpython==0.14.2
configparser==3.5.0b2
curtsies==0.1.19
dj-database-url==0.3.0
dj-static==0.0.6
django-allauth==0.19.1
django-extensions==1.5.5
django-reversion==1.8.7
django-toolbelt==0.0.1
djangorestframework==3.1.3
env-tools==2.0.0
greenlet==0.4.7
gunicorn==19.3.0
oauthlib==0.7.2
psycopg2==2.6
python-openid==2.2.5
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
static3==0.6.1
tini==3.0.1
wsgiref==0.1.2
